etl/requirements.txt
asgiref==3.7.1
cmake==3.26.4
coverage==7.2.6
django==4.2.2
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-split-settings==1.2.0
exceptiongroup==1.1.1
Faker==18.9.0
flake8==4.0.0
iniconfig==2.0.0
mccabe==0.6.1
now==0.0.6
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==7.3.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python_dotenv==1.0.0
pytz==2023.3
six==1.16.0
split-settings==1.0.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.1
uwsgi
flake8==4.0  # Разумеется, вам потребуется линтер :)
django-cors-headers==4.1.0
psycopg==3.1.9
psycopg-binary==3.1.9
pydantic==1.10.9

backoff
requests
elasticsearch==8.5.0
